+++ test/integration/backends/my_module.py
@@ -1,6 +1,7 @@
+import time
+
 from superduper import logging
 
-import time
 
 def my_func(x):
     logging.info(f'my_func({x}) executing')
@@ -14,5 +15,3 @@ def my_other_func(x):
     time.sleep(1)
     logging.info(f'my_other_func({x}) executing... DONE')
     return x
-
-