+++ test/integration/backends/test_simple.py
@@ -1,8 +1,7 @@
-from superduper import *
-from superduper.base.event import Job
-
 from my_module import my_func, my_other_func
 
+from superduper import ObjectModel, superduper
+from superduper.base.event import Job
 
 
 def test():